Migrate availabilityBlobDescription logic into an availabilityService
